+++ mayan/apps/metadata/events.py
@@ -6,6 +6,21 @@ from events import EventTypeNamespace
 
 namespace = EventTypeNamespace(name='metadata', label=_('Metadata'))
 
+event_document_metadata_added = namespace.add_event_type(
+    name='document_metadata_added', label=_(
+        'Document metadata added'
+    )
+)
+event_document_metadata_edited = namespace.add_event_type(
+    name='document_metadata_edited', label=_(
+        'Document metadata edited'
+    )
+)
+event_document_metadata_removed = namespace.add_event_type(
+    name='document_metadata_removed', label=_(
+        'Document metadata removed'
+    )
+)
 event_metadata_type_created = namespace.add_event_type(
     name='metadata_type_created', label=_('Metadata type created')
 )
@@ -17,3 +32,4 @@ event_metadata_type_relationship = namespace.add_event_type(
         'Metadata type relationship updated'
     )
 )
+